+++ frontend/public/components/utils/index.tsx
@@ -52,7 +52,7 @@ export * from './poll-hook';
 export * from './ref-width-hook';
 export * from './safe-fetch-hook';
 export * from './camel-case-wrap';
-export * from './truncate';
+export * from './truncate-middle';
 
 /*
   Add the enum for NameValueEditorPair here and not in its namesake file because the editor should always be
